+++ src/main/java/com/synopsys/integration/detect/configuration/DetectProperties.java
@@ -51,7 +51,6 @@ import com.synopsys.integration.detect.configuration.enumeration.RapidCompareMod
 import com.synopsys.integration.detect.tool.signaturescanner.enums.ExtendedIndividualFileMatchingMode;
 import com.synopsys.integration.detect.tool.signaturescanner.enums.ExtendedSnippetMode;
 import com.synopsys.integration.detect.workflow.DummyAccuracyEnum;
-import com.synopsys.integration.detect.workflow.bdio.AggregateMode;
 import com.synopsys.integration.detectable.detectables.bazel.WorkspaceRule;
 import com.synopsys.integration.detectable.detectables.bitbake.BitbakeDependencyType;
 import com.synopsys.integration.detectable.detectables.conan.cli.config.ConanDependencyType;
@@ -1737,7 +1736,6 @@ public class DetectProperties {
 
     //#region Deprecated Properties
     // username/password ==> api token
-    private static final String BDIO1_DEPRECATION_MESSAGE = "This property is being removed, along with the option to generate BDIO in BDIO1 format. In the future, BDIO2 format will be the only option.";
     private static final String AGGREGATION_MODE_DEPRECATION_MESSAGE = "This property is being removed, along with the ability to set the aggregation mode. In the future, Detect will always operate in SUBPROJECT aggregation mode (regardless of how it is configured) to more accurately report the dependency graph.";
     private static final String BAZEL_DEPENDENCY_TYPE_DEPRECATION_MESSAGE = "This property is being removed. Please use property 'detect.bazel.workspace.rules' instead.";
     private static final String DETECT_DOCKER_PATH_REQUIRED_DEPRECATION_MESSAGE = "This property is being removed. A docker executable is only required when running the Docker tool in air gap mode.";
@@ -1761,47 +1759,6 @@ public class DetectProperties {
             .setGroups(DetectGroup.BAZEL, DetectGroup.SOURCE_SCAN)
             .setDeprecated(BAZEL_DEPENDENCY_TYPE_DEPRECATION_MESSAGE, DetectMajorVersion.EIGHT)
             .build();
-
-    @Deprecated
-    public static final NullableStringProperty DETECT_BOM_AGGREGATE_NAME =
-        NullableStringProperty.newBuilder("detect.bom.aggregate.name")
-            .setInfo("Aggregate BDIO File Name", DetectPropertyFromVersion.VERSION_3_0_0)
-            .setHelp("If set, this will aggregate all the BOMs to create a single BDIO file with the filename provided.")
-            .setGroups(DetectGroup.PROJECT, DetectGroup.PROJECT_SETTING)
-            .setCategory(DetectCategory.Advanced)
-            .setDeprecated(
-                "This property is being removed. Use detect.bdio.file.name to control the name of the bdio file Detect generates. Currently detect.bdio.file.name has the same effects as this property. In the future, Detect will always operate in SUBPROJECT aggregation mode regardless of how it is configured; detect.bdio.file.name will only control the BDIO file name.",
-                DetectMajorVersion.EIGHT
-            )
-            .build();
-
-    @Deprecated
-    public static final EnumProperty<AggregateMode> DETECT_BOM_AGGREGATE_REMEDIATION_MODE =
-        EnumProperty.newBuilder("detect.bom.aggregate.remediation.mode", AggregateMode.TRANSITIVE, AggregateMode.class)
-            .setInfo("BDIO Aggregate Remediation Mode", DetectPropertyFromVersion.VERSION_6_1_0)
-            .setHelp(
-                "If an aggregate BDIO file is being generated (that is, property detect.bom.aggregate.name has been set) " +
-                    "and this property is set to DIRECT, the aggregate BDIO file will exclude code location nodes " +
-                    "from the top layer of the dependency tree to preserve the correct identification of direct dependencies in the resulting Black Duck BOM. " +
-                    "When this property is set to TRANSITIVE (the default), component source information is preserved by including code location nodes at the " +
-                    "top of the dependency tree, but all components will appear as TRANSITIVE in the BOM. " +
-                    "SUBPROJECT aggregation mode provides both component source information and correct identification of direct and transitive dependencies by " +
-                    "encoding code location nodes as subprojects in the graph. SUBPROJECT aggregation mode must only be used with Black Duck 2021.8.0 or later, " +
-                    "and has no effect (is equivalent to TRANSITIVE mode) when detect.bdio2.enabled is set to false.")
-            .setGroups(DetectGroup.PROJECT, DetectGroup.PROJECT_SETTING)
-            .setCategory(DetectCategory.Advanced)
-            .setDeprecated(AGGREGATION_MODE_DEPRECATION_MESSAGE, DetectMajorVersion.EIGHT)
-            .build();
-
-    @Deprecated
-    public static final BooleanProperty DETECT_BDIO2_ENABLED =
-        BooleanProperty.newBuilder("detect.bdio2.enabled", true)
-            .setInfo("BDIO 2 Enabled", DetectPropertyFromVersion.VERSION_6_1_0)
-            .setHelp("The version of BDIO files to generate.", "If set to false, BDIO version 1 will be generated. If set to true, BDIO version 2 will be generated.")
-            .setGroups(DetectGroup.PATHS, DetectGroup.GLOBAL)
-            .setDeprecated(BDIO1_DEPRECATION_MESSAGE, DetectMajorVersion.EIGHT)
-            .build();
-
     @Deprecated
     public static final BooleanProperty DETECT_BUILDLESS =
         BooleanProperty.newBuilder("detect.detector.buildless", false)