+++ src/main/java/com/synopsys/integration/detect/tool/impactanalysis/ImpactAnalysisNamingOperation.java
@@ -1,7 +1,6 @@
 package com.synopsys.integration.detect.tool.impactanalysis;
 
 import java.io.File;
-import java.io.IOException;
 
 import com.synopsys.integration.detect.workflow.codelocation.CodeLocationNameManager;
 import com.synopsys.integration.util.NameVersion;
@@ -13,12 +12,10 @@ public class ImpactAnalysisNamingOperation {
         this.codeLocationNameManager = codeLocationNameManager;
     }
 
-    public String createCodeLocationName(File toScan, NameVersion projectNameAndVersion, ImpactAnalysisOptions impactAnalysisOptions) throws IOException {
+    public String createCodeLocationName(File toScan, NameVersion projectNameAndVersion) {
         String projectName = projectNameAndVersion.getName();
         String projectVersionName = projectNameAndVersion.getVersion();
-        String codeLocationPrefix = impactAnalysisOptions.getCodeLocationPrefix();
-        String codeLocationSuffix = impactAnalysisOptions.getCodeLocationSuffix();
-        return codeLocationNameManager.createImpactAnalysisCodeLocationName(toScan, projectName, projectVersionName, codeLocationPrefix, codeLocationSuffix);
+        return codeLocationNameManager.createImpactAnalysisCodeLocationName(toScan, projectName, projectVersionName);
     }
 
 }